pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-studio"
version = "0.1.0"
description = "Agent Studio"
requires-python = ">=3.11"
dependencies = [
    "watchdog>=3.0.0",
    "pyyaml>=6.0",
    "fasteners>=0.19",
    "websockets>=12.0",
    "aiohttp>=3.9",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src", "src.*"]

[project.scripts]
agent-studio = "src.cli.main:main"

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
addopts = "--import-mode=importlib -p anyio"

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-anyio>=0.0.0",
]
